requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.2
schedule==1.2.0
